requirements.txt
Django
google-api-python-client
requests
pytrends